Fix ZeroMQ announce CI gates
